agent/requirements.txt
# FastAPI and web server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# LangGraph OSS (no licensing required)
langgraph>=0.2.0
langgraph-checkpoint-postgres>=0.1.0

# HTTP client for Ollama
requests>=2.31.0

# PostgreSQL support
psycopg2-binary>=2.9.0

# Logging and utilities
python-multipart>=0.0.6
pydantic>=2.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
